backend/requirements.txt
beautifulsoup4==4.12.3
fastapi==0.112.0
httpx==0.27.0
numpy==1.23.4
openai==1.40.1
passlib==1.7.4
pydantic==2.8.2
PyJWT==2.9.0
pytest==8.3.2
pytest_asyncio==0.23.8
python-dotenv==1.0.1
python-multipart==0.0.9
Requests==2.32.3
scikit_learn==1.2.0
selenium==4.23.1
SQLAlchemy==2.0.31
SQLAlchemy_Utils==0.41.2
yfinance==0.2.41
uvicorn == 0.30.5
PyMySql==1.1.1
# passlib[bcrypt]==1.7.4